requirements.txt
fastapi
uvicorn
openai
python-dotenv
faiss-cpu
PyMuPDF
python-docx
streamlit
langchain
tqdm